adding changes to jumbotron and card-list for alight
