+++ blocks/init/src/blocks/custom/card-list/card-list-block.js
@@ -1,10 +1,11 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
 import { Fragment } from '@wordpress/element';
-import { InspectorControls } from '@wordpress/block-editor';
+import { InspectorControls, BlockControls } from '@wordpress/block-editor';
 import { getActions } from '@eightshift/frontend-libs/scripts/editor';
 import manifest from './manifest.json';
 import { CardListEditor } from './components/card-list-editor';
 import { CardListOptions } from './components/card-list-options';
+import { CardListToolbar } from './components/card-list-toolbar';
 
 export const CardList = (props) => {
 
@@ -22,6 +23,12 @@ export const CardList = (props) => {
           actions={actions}
         />
       </InspectorControls>
+      <BlockControls>
+        <CardListToolbar
+          attributes={attributes}
+          actions={actions}
+        />
+      </BlockControls>
       <CardListEditor
         attributes={attributes}
         actions={actions}
